Abstract the ringing feature to make it easier to use (#5390)
