+++ lua/sim/commands/shared.lua
@@ -357,56 +357,3 @@ function FindNearestUnit(units, px, pz)
 
     return nearest
 end
-
--- Scan and gather build skirts in the surrounding area. This is used as an alternative 
--- to relying on `brain:CanBuildStructureAt(...)` as it returns false positives.
----@param cx number
----@param cz number
----@param skirtSize number
----@param cx1 number[]  # re-useable array
----@param cz1 number[]  # re-useable array
----@param cx2 number[]  # re-useable array
----@param cz2 number[]  # re-useable array
----@return number[]
----@return number[]
----@return number[]
----@return number[]
----@return number
-function FindBuildingSkirts(cx, cz, skirtSize, cx1, cz1, cx2, cz2)
-
-    local x1 = cx - (skirtSize + 10)
-    local z1 = cz - (skirtSize + 10)
-    local x2 = cx + (skirtSize + 10)
-    local z2 = cz + (skirtSize + 10)
-
-    -- clear out the cache
-    for k = 1, table.getn(cx1) do
-        cx1[k] = nil
-        cz1[k] = nil
-        cx2[k] = nil
-        cz2[k] = nil
-    end
-
-    -- find all units that may prevent us from building
-    local structures = GetUnitsInRect(x1, z1, x2, z2)
-    if not structures then
-        return cx1, cz1, cx2, cz2, 0
-    end
-
-    structures = EntityCategoryFilterDown(categories.STRUCTURE + categories.EXPERIMENTAL, structures)
-
-    -- populate the skirts to check
-    local buildingSkirtCount = 0
-    for k, unit in structures do
-        local blueprint = unit:GetBlueprint()
-        local px, _, pz = unit:GetPositionXYZ()
-        local sx, sz = 0.5 * blueprint.Physics.SkirtSizeX, 0.5 * blueprint.Physics.SkirtSizeZ
-        cx1[k] = px - sx
-        cz1[k] = pz - sz
-        cx2[k] = px + sz
-        cz2[k] = pz + sz
-        buildingSkirtCount = buildingSkirtCount + 1
-    end
-
-    return cx1, cz1, cx2, cz2, buildingSkirtCount
-end
